pyproject.toml
[project]
name = "vibe-coding"
version = "0.1.0"
description = "Coze Vibe Coding project"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.0.3",
    "langchain-openai==1.0.1",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.0",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.39",
    "fastapi>=0.121,<1",
    "uvicorn>=0.38,<1",
    "pydantic>=2.12,<3",
    "Jinja2>=3.1,<4",
    "coze-coding-utils>=0.2.6,<1",
    "coze-workload-identity>=0.1.4,<1",
    "opencv-python>=4.12,<5",
    "pandas>=2.2,<3",
    "pillow>=10.3,<11",
    "cozeloop>=0.1.25,<1",
    "alembic>=1.16,<2",
    "SQLAlchemy>=2.0,<3",
    "python-dotenv>=1.2,<2",
    "psycopg2-binary>=2.9,<3",
    "boto3>=1.40,<2",
    "langgraph-checkpoint-postgres>=3.0,<4",
    "psycopg[binary]>=3.3,<4",
    "psycopg-pool>=3.3,<4",
    "pypdf>=6.4,<7",
    "docx2python>=3.5,<4",
    "openpyxl>=3.1,<4",
    "python-pptx>=1.0,<2",
    "coze-coding-dev-sdk>0.5.0,<1",
    "rich>=14,<15",
    "chardet>=5.2,<6",
    "pycairo==1.29.0",
    "dbus-python==1.3.2",
    "PyGObject==3.48.2",
    "requests>=2.32.5",
]

[tool.uv]
# 使用阿里云镜像作为默认源，替代 pypi.org
[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
